requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
kombu==5.2.3
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.22.1
packaging==21.3
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
redis==4.1.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
threadpoolctl==3.0.0
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.7.0
